+++ client-modern/src/main/java/com/paraguacraft/pvp/modern/mixin/HandledScreenAccessor.java
@@ -12,4 +12,7 @@ public interface HandledScreenAccessor {
 
     @Accessor("y")
     int paraguacraft$getGuiY();
+
+    @Accessor("backgroundWidth")
+    int paraguacraft$getBackgroundWidth();
 }